Changing the variables used for storing buffer lengths and file offsets for parallel file IO to be double precision since it happens that for longer runs it might reach sizes which exceed that of single precision.
git-svn-id: svn+ssh://svn.wcas.northwestern.edu/var/svn/cmc/branches/cmc-mpi@1048 46abec54-52ec-0310-8d7b-a975267b85f2

+++ cmc.h
@@ -1364,7 +1364,7 @@ void ComputeEnergy(void);
 
 #ifdef USE_MPI
 void mpi_ComputeEnergy(void);
-void mpi_para_file_write(char* wrbuf, int *len, int *prev_cum_offset, MPI_File* fh);
+void mpi_para_file_write(char* wrbuf, long long *len, long long *prev_cum_offset, MPI_File* fh);
 void PrintParaFileOutput(void);
 void mpi_close_node_buffers(void);
 #endif
